ref: Destructurised filters url handling to separate hook
